Finished solver and istancebenchmark c++.
Next step is implementing the combinator
and brute force.

The portfolio should then be implemented,
and then this would be a nice scaffolding
for implementing AMF techniques.

These techniques should also be implemented
as OpenTuner techniques and metatechniques.

+++ cpp-combinator/lib/InstanceBenchmark.cpp
@@ -2,7 +2,7 @@
 
 InstanceBenchmark::InstanceBenchmark(std::string new_instance, 
                                      std::string new_solver,
-                                     float* new_values) {
+                                     double* new_values) {
     solver   = new_solver;
     instance = new_instance;
     values   = new_values;
@@ -16,24 +16,24 @@ std::string InstanceBenchmark::get_instance_name() {
 std::string InstanceBenchmark::get_solver_name() {
     return solver;
 }
-float InstanceBenchmark::get_mean() {
+double InstanceBenchmark::get_mean() {
     return mean;
 }
-float* InstanceBenchmark::get_values() {
+double* InstanceBenchmark::get_values() {
     return values;
 }
-float InstanceBenchmark::get_median() {
+double InstanceBenchmark::get_median() {
     return median;
 }
-float InstanceBenchmark::get_stddev() {
+double InstanceBenchmark::get_stddev() {
     return stddev;
 }
-float InstanceBenchmark::calc_stddev() {
+double InstanceBenchmark::calc_stddev() {
     return 0.0;
 }
-float InstanceBenchmark::calc_mean() {
+double InstanceBenchmark::calc_mean() {
     return 0.0;
 }
-float InstanceBenchmark::calc_median() {
+double InstanceBenchmark::calc_median() {
     return 0.0;
 }
